Fix extracting triples maps in RML
Checking on "<TM> a rr:TriplesMap" is not sufficient since this	triple
might not exist. New RML requires the triples map to have a subject map
and old RML also requires a logical source. This fix check on having a subject
map for	both versions of RML. A	logical	source is not checked since AlgeMapLoom
will check that	later on.